Applied new Architecture, new one is easier to work with in terms of multi tier architectures
