--- .idea/compiler.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <annotationProcessing>
-      <profile name="Maven default annotation processors profile" enabled="true">
-        <sourceOutputDir name="target/generated-sources/annotations" />
-        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
-        <outputRelativeToContentRoot value="true" />
-        <module name="MCLauncher.API" />
-        <module name="MCLauncher.BLL" />
-        <module name="MCLauncher.DAL" />
-        <module name="MCLauncher.Domain" />
-        <module name="MCLauncher.UI.Admin" />
-        <module name="MCLauncher.UI.EndUser" />
-        <module name="MCLauncherAPI" />
-        <module name="MCLauncherBLL" />
-        <module name="MCLauncherDAL" />
-        <module name="MCLauncherDomain" />
-        <module name="MCLauncherUIAdmin" />
-        <module name="MCLauncherUIUser" />
-      </profile>
-    </annotationProcessing>
-  </component>
-</project>
