+++ MCLEndUser/src/main/java/net/toastynetworks/MCLEndUser/Controller.java
@@ -1,0 +1,8 @@
+package net.toastynetworks.MCLEndUser;
+
+public class Controller {
+
+    public void playButtonClicked(){
+        System.out.println("Used started modpack....");
+    }
+}
