+++ MCLEndUser/src/main/resources/user-ui.fxml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.Cursor?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ContextMenu?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.MenuItem?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="653.0" style="-fx-background-color: #1a1a1a;" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="net.toastynetworks.MCLEndUser.Controller">
+    <children>
+        <ListView layoutX="8.0" layoutY="88.0" prefHeight="312.0" prefWidth="263.0" style="-fx-background-color: #e8b23e;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.topAnchor="88.0">
+            <contextMenu>
+                <ContextMenu>
+                    <items>
+                        <MenuItem mnemonicParsing="false" text="Launch" />
+                        <MenuItem mnemonicParsing="false" text="Settings" />
+                        <MenuItem mnemonicParsing="false" text="Uninstall" />
+                        <MenuItem mnemonicParsing="false" text="Screenshots" />
+                        <MenuItem mnemonicParsing="false" text="Logs" />
+                        <MenuItem mnemonicParsing="false" text="Crashlogs" />
+                        <MenuItem mnemonicParsing="false" text="Modpack Folder" />
+                    </items>
+                </ContextMenu>
+            </contextMenu>
+        </ListView>
+        <Pane layoutY="-1.0" prefHeight="90.0" prefWidth="653.0" style="-fx-background-color: #1a1a1a;" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
+            <children>
+                <ImageView fitHeight="86.0" fitWidth="283.0" layoutX="1.0" pickOnBounds="true" preserveRatio="true">
+                    <image>
+                        <Image url="https://toastynetworks.net/graphics/Logo_WithGears.png" />
+                    </image>
+                </ImageView>
+            </children>
+        </Pane>
+        <Pane prefHeight="90.0" prefWidth="138.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+            <children>
+                <ImageView fitHeight="63.0" fitWidth="45.0" layoutX="7.0" layoutY="26.0" pickOnBounds="true" preserveRatio="true">
+                    <image>
+                        <Image url="https://toastynetworks.net/graphics/avatar.png" />
+                    </image>
+                </ImageView>
+                <Label layoutX="63.0" layoutY="38.0" prefHeight="20.0" prefWidth="68.0" text="Rubbertjuh" textFill="WHITE">
+                    <font>
+                        <Font name="NexaBold" size="12.0" />
+                    </font>
+                </Label>
+            </children>
+        </Pane>
+        <Pane layoutX="264.0" layoutY="88.0" prefHeight="312.0" prefWidth="389.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="264.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="88.0">
+            <children>
+                <ImageView fitHeight="1026.0" fitWidth="1923.0" layoutX="-1.0" opacity="0.52" pickOnBounds="true">
+                    <image>
+                        <Image url="https://toastynetworks.net/graphics/beyond-full.png" />
+                    </image>
+                </ImageView>
+            </children>
+        </Pane>
+        <Pane layoutX="216.0" layoutY="10.0" prefHeight="90.0" prefWidth="315.0">
+            <children>
+                <Text fill="WHITE" layoutX="21.0" layoutY="50.0" strokeType="OUTSIDE" strokeWidth="0.0" text="MODPACKS" wrappingWidth="75.00000366568565">
+                    <font>
+                        <Font name="NexaBold" size="12.0" />
+                    </font>
+                </Text>
+                <Text fill="WHITE" layoutX="113.0" layoutY="50.0" strokeType="OUTSIDE" strokeWidth="0.0" text="NEWS" wrappingWidth="44.99998325109482">
+                    <font>
+                        <Font name="NexaBold" size="12.0" />
+                    </font>
+                </Text>
+                <Text fill="#fcfcfc" layoutX="169.0" layoutY="50.0" strokeType="OUTSIDE" strokeWidth="0.0" text="STORE" wrappingWidth="44.99998325109482">
+                    <font>
+                        <Font name="NexaBold" size="12.0" />
+                    </font>
+                </Text>
+                <Text fill="WHITE" layoutX="224.0" layoutY="50.0" strokeType="OUTSIDE" strokeWidth="0.0" text="SUPPORT" wrappingWidth="68.00000366568565">
+                    <font>
+                        <Font name="NexaBold" size="12.0" />
+                    </font>
+                </Text>
+            </children>
+        </Pane>
+        <Button layoutX="458.0" layoutY="330.0" mnemonicParsing="false" onAction="#playButtonClicked" prefHeight="56.0" prefWidth="126.0" style="-fx-background-color: #e8b23e; -fx-background-radius: 5px;" text="PLAY" textFill="WHITE" AnchorPane.bottomAnchor="15.0" AnchorPane.rightAnchor="30.0">
+            <font>
+                <Font name="NexaBold" size="18.0" />
+            </font>
+            <cursor>
+                <Cursor fx:constant="HAND" />
+            </cursor>
+        </Button>
+    </children>
+</AnchorPane>
