+++ MCLRestAPI/src/main/java/net/toastynetworks/MCLRestAPI/Application.java
@@ -1,0 +1,14 @@
+package net.toastynetworks.MCLRestAPI;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@EnableSwagger2
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
