--- MCLauncher.API/src/main/Java/net/toastynetworks/MCLauncher/API/Application.java
@@ -1,14 +1,0 @@
-package net.toastynetworks.MCLauncher.API;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@SpringBootApplication
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-}
