--- MCLauncher.API/src/test/java/net/toastynetworks/MCLauncher/API/ApplicationTests.java
@@ -1,16 +1,0 @@
-package net.toastynetworks.MCLauncher.API;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
