--- MCLauncher.DAL/src/main/Java/net/toastynetworks/MCLauncher/DAL/ContextInterfaces/IModpackContext.java
@@ -1,4 +1,0 @@
-package net.toastynetworks.MCLauncher.DAL.ContextInterfaces;
-
-public interface IModpackContext {
-}
