--- MCLauncher.DAL/src/main/Java/net/toastynetworks/MCLauncher/DAL/Contexts/ModpackMySQLContext.java
@@ -1,4 +1,0 @@
-package net.toastynetworks.MCLauncher.DAL.Contexts;
-
-public class ModpackMySQLContext {
-}
