--- MCLauncher.DAL/src/main/Java/net/toastynetworks/MCLauncher/DAL/Repositories/ModpackRepository.java
@@ -1,4 +1,0 @@
-package net.toastynetworks.MCLauncher.DAL.Repositories;
-
-public class ModpackRepository {
-}
