--- MCLauncher.DAL/src/main/Java/net/toastynetworks/MCLauncher/DAL/RepositoryInterfaces/IModpackRepository.java
@@ -1,4 +1,0 @@
-package net.toastynetworks.MCLauncher.DAL.RepositoryInterfaces;
-
-public interface IModpackRepository {
-}
