--- MCLauncher.UI.EndUser/src/main/Java/net/toastynetworks/MCLauncher/UI/EndUser/Controller.java
@@ -1,8 +1,0 @@
-package net.toastynetworks.MCLauncher.UI.EndUser;
-
-public class Controller {
-
-    public void playButtonClicked(){
-        System.out.println("Used started modpack....");
-    }
-}
